+++ CHANGELOG.md
@@ -1,9 +1,8 @@
-Unreleased
+Development version
 ----------------------------------------
-
 - Fixed bug for gradient norm calculation inside `ControlModel`
 - Added support for generic risk measures and chance constrained optimization through `TransformedMeanRiskMeasureSAA` and `PenalizationControlcostFunctional`
-
+- Added jupyter notebook tutorials 
 
 Version 1.0.0, released on Aug 11, 2023
 ----------------------------------------